transfert opcodes and test
