--- cpu/mos6502/src/tsx.c
@@ -1,8 +1,0 @@
-#include <mos6502.h>
-
-static int tsx(mos6502_t *cpu)
-{
-    unsigned short address = cpu->read16(cpu, cpu->pc++);
-    cpu->x &= address + cpu->sp;
-    return 2;
-}
