--- cpu/mos6502/test/tay.c
@@ -1,16 +1,0 @@
-#include "test.h"
-
-static int test_tay()
-{
-    unsigned char rom = 0xA8;
-
-    mos6502_t cpu;
-    mos6502_init(&cpu);
-    mos6502_add_test_full_mapping(&cpu, rom);
-
-
-    assert_is_equal(ticks, 2);
-    assert_is_equal(cpu.y, cpu.a);
-
-    return 0;
-}
